Fixed server bugs

+++ src/io/device_serial_port.cpp
@@ -59,7 +59,7 @@ void DeviceSerialPort::Gather()
                 currentDevice = (Device)(b[0] >> 7);
 
                 // Bits 4-6 will tell us how many extra bytes are coming from this device.
-                bytesForCurrent = 1 + ((int)b[0] & 0b01110000) >> 4;
+                bytesForCurrent = 1 + ((b[0] & 0b01110000) >> 4);
             }
